Refactored the models to a structure that makes more sense.
