+++ container_pipeline/models/tracking.py
@@ -1,7 +1,4 @@
-import subprocess
-
 from django.db import models
-from django.conf import settings
 
 
 class Package(models.Model):
@@ -40,72 +37,3 @@ class RepoInfo(models.Model):
         return self.baseurls
 
 
-class ContainerImage(models.Model):
-    name = models.CharField(max_length=200, db_index=True,
-                            help_text="Image name", unique=True)
-    packages = models.ManyToManyField(Package, related_name='images',
-                                      help_text="Packages")
-    parents = models.ManyToManyField('self', symmetrical=False,
-                                     help_text="Parent images")
-    repoinfo = models.ForeignKey(RepoInfo, null=True)
-
-    to_build = models.BooleanField(default=False, db_index=True,
-                                   help_text='Whether to build image or not')
-    scanned = models.BooleanField(default=False, db_index=True,
-                                  help_text='Whether the image is scanner')
-    last_scanned = models.DateTimeField(null=True, blank=True)
-
-    created = models.DateTimeField(auto_now_add=True, blank=True)
-    last_updated = models.DateTimeField(auto_now=True, blank=True)
-
-    class Meta:
-        db_table = 'container_images'
-        app_label = 'container_pipeline'
-
-    def __str__(self):
-        return self.name
-
-    @property
-    def fullname(self):
-        return '{}/{}'.format(settings.REGISTRY_ENDPOINT[0], self.name)
-
-    @property
-    def jobname(self):
-        return self.name.replace(':', '-').replace('/', '-')
-
-    def pull(self):
-        docker_cmd = 'docker pull {name}'.format(name=self.fullname)
-        return subprocess.check_output(docker_cmd, shell=True)
-
-    def run(self, cmd):
-        docker_cmd = (
-            'docker run --rm --entrypoint "/bin/bash" {name} -c "{cmd}"'.
-            format(name=self.fullname, cmd=cmd.replace('"', '\\"')))
-        return subprocess.check_output(docker_cmd, shell=True)
-
-    def remove(self):
-        """Remove docker image"""
-        docker_cmd = (
-            'docker rmi -f {}'.format(self.fullname))
-        return subprocess.check_output(docker_cmd, shell=True)
-
-    def trigger_build(self):
-        if settings.JENKINS_USERNAME:
-            cmd = (
-                'java -jar {} -s {} build {} --username {} --password {}'
-                .format(
-                    settings.JENKINS_CLI,
-                    settings.JENKINS_ENDPOINT,
-                    self.jobname,
-                    settings.JENKINS_USERNAME,
-                    settings.JENKINS_PASSWORD)
-            )
-        else:
-            cmd = (
-                'java -jar {} -s {} build {}'.format(
-                    settings.JENKINS_CLI,
-                    settings.JENKINS_ENDPOINT,
-                    self.jobname)
-            )
-
-        subprocess.check_call(cmd, shell=True)
